+++ CHANGES.rst
@@ -5,12 +5,12 @@ Changelog
 4.2rc16 (unreleased)
 --------------------
 
+- Fixed `searchitemstoprevalidate` collection TAL condition,
+  state is `prevalidated` not `pre_validated`.
+  [gbastien]
 - Added WFAdaptation `item_validation_shortcuts` that will let users change item
   state to any other item validation state (so between itemcreated and validated)
   depending on their groups.
-  [gbastien]
-- Fixed `searchitemstoprevalidate` collection TAL condition,
-  state is `prevalidated` not `pre_validated`.
   [gbastien]
 
 4.2rc15 (2022-02-25)